feat(tooling): replaced poetry with uv and added RUFF support.
Also added formatting changes but didn't rectify issues related to RUFF, MYPY or Pyright
