+++ .editorconfig
@@ -1,0 +1,23 @@
+root = true
+
+[*]
+tab_width = 4
+end_of_line = lf
+max_line_length = 120
+ij_visual_guides = 120
+insert_final_newline = true
+trim_trailing_whitespace = true
+
+[*.md]
+trim_trailing_whitespace = false
+
+[*.{yml,yaml}]
+indent_style = tab
+indent_size = 4
+
+
+[*.py]
+indent_style = tab
+indent_size = 4
+ij_python_from_import_parentheses_force_if_multiline = true
+charset = utf-8
