+++ pyproject.toml
@@ -1,41 +1,71 @@
-[tool.poetry]
-name = "query_generator"
+[project]
+name = "python-query-generator"
 version = "0.1.0"
 description = "A library for generating queries based on a central JSON configuration"
-authors = ["Bhavani Ravi <bhavanicodes@gmail.com>"]
-readme = "README.md"
 license = "Apache-2.0"
+readme = "README.md"
+requires-python = ">=3.12"
+authors = [{ name = "Bhavani Ravi", email = "bhavanicodes@gmail.com" }]
+dependencies = [
+    "pydantic>=2.9.2",
+    "pydantic-settings>=2.6.0",
+    "sqlmodel>=0.0.22",
+    "sqlparse>=0.5.1",
+    "hatchling>=1.25.0",
+]
 
-[tool.poetry.dependencies]
-python = ">=3.9,<=3.12.*"
-sqlparse = "^0.5.1"
-pydantic = "^2.9.2"
-sqlmodel = "^0.0.22"
-pydantic-settings = "^2.6.0"
+[project.optional-dependencies]
+mongodb = [
+    "pymongo>=4.10.1",
+]
+bigquery = [
+    "google-cloud-bigquery>=3.26.0",
+]
+postgresql = [
+    "psycopg>=3.2.3",
+]
 
-[tool.poetry.group.postgresql.dependencies]
-psycopg2 = "^2.9.10"
-
-[tool.poetry.group.mongodb.dependencies]
-pymongo = "^4.10.1"
-
-[tool.poetry.group.bigquery.dependencies]
-google-cloud-bigquery = "^3.26.0"
-
-
-
-[tool.poetry.group.dev.dependencies]
-ruff = "^0.7.0"
-
-[tool.poetry.dev-dependencies]
-black = "^24.10.0"
-flake8 = "^7.1.1"
-isort = "^5.13.2"
-mypy = "^1.12.0"
-pytest = "^8.3.3"
-pytest-cov = "^5.0.0"
+[tool.uv]
+dev-dependencies = [
+    "mypy>=1.13.0",
+    "pytest>=8.3.3",
+    "pytest-cov>=5.0.0",
+    "pytest-richer>=0.1.1",
+    "ruff>=0.7.1",
+    "pyright>=1.1.386",
+]
 
 
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["query_generator", "scripts"] # Support flat layout
+
+[tool.pyright]
+exclude = ["tests", ".venv"]
+
+[tool.coverage.run]
+omit = [
+    "tests/*"
+]
+relative_files = true
+branch = true
+
+[tool.coverage.report]
+exclude_lines = [
+    "pragma: no cover",
+    "class .*\\(.*\\bProtocol\\b.*\\):",
+    "@(?:typing\\.)?overload",
+    "if TYPE_CHECKING:",
+    "raise NotImplementedError"
+]
+omit = [
+    "tests/*",
+]
+show_missing = true
+skip_covered = true
+skip_empty = true
+precision = 2
+fail_under = 80
