+++ query_generator/schemas.py
@@ -1,5 +1,6 @@
-from pydantic import BaseModel, Field
 from typing import Any, Dict
+
+from pydantic import BaseModel, Field
 
 
 class ConditionEnum(str):
