+++ query_generator/utils/connections/base_connection.py
@@ -2,6 +2,6 @@ from abc import ABC, abstractmethod
 
 
 class BaseConnection(ABC):
-    @abstractmethod
-    def connect(self):
-        pass
+	@abstractmethod
+	def connect(self):
+		pass
