+++ query_generator/utils/query_generators/base_query_generator.py
@@ -2,6 +2,6 @@ from abc import ABC, abstractmethod
 
 
 class BaseQueryGenerator(ABC):
-    @abstractmethod
-    def generate_query(self, config):
-        pass
+	@abstractmethod
+	def generate_query(self, config):
+		pass
