--- requirements.txt
@@ -1,53 +1,0 @@
-annotated-types==0.7.0
-black==24.10.0
-cachetools==5.5.0
-certifi==2024.8.30
-charset-normalizer==3.4.0
-click==8.1.7
-coverage==7.6.3
-dnspython==2.7.0
-flake8==7.1.1
-google-api-core==2.21.0
-google-auth==2.35.0
-google-cloud-bigquery==3.26.0
-google-cloud-core==2.4.1
-google-crc32c==1.6.0
-google-resumable-media==2.7.2
-googleapis-common-protos==1.65.0
-grpcio==1.67.0
-grpcio-status==1.67.0
-idna==3.10
-iniconfig==2.0.0
-isort==5.13.2
-mccabe==0.7.0
-mypy==1.12.0
-mypy-extensions==1.0.0
-packaging==24.1
-pathspec==0.12.1
-platformdirs==4.3.6
-pluggy==1.5.0
-proto-plus==1.24.0
-protobuf==5.28.2
-psycopg2==2.9.10
-pyasn1==0.6.1
-pyasn1_modules==0.4.1
-pycodestyle==2.12.1
-pydantic==2.9.2
-pydantic-settings==2.6.0
-pydantic_core==2.23.4
-pyflakes==3.2.0
-pymongo==4.10.1
-pytest==8.3.3
-pytest-cov==5.0.0
-python-dateutil==2.9.0.post0
-python-dotenv==1.0.1
--e git+ssh://git@github.com/bhavaniravi/python-query-generator.git@2d5a559ce96a8a98c7d33085570acb348bf77c2f#egg=query_generator
-requests==2.32.3
-rsa==4.9
-ruff==0.7.0
-six==1.16.0
-SQLAlchemy==2.0.36
-sqlmodel==0.0.22
-sqlparse==0.5.1
-typing_extensions==4.12.2
-urllib3==2.2.3
