+++ scripts/generate_json_schema.py
@@ -1,6 +1,7 @@
-from query_generator.schemas import QueryConfig
 import json
 from pathlib import Path
+
+from query_generator.schemas import QueryConfig
 
 main_model_schema = QueryConfig.model_json_schema()
 out = json.dumps(main_model_schema, indent=2)
@@ -9,4 +10,4 @@ project_path = Path(__file__).parent.parent
 file_path = project_path / "static" / "query_json_schema.json"
 
 with open(file_path, "w") as f:
-    f.write(out)
+	f.write(out)
